changed homepage from stateful to getview
