+++ with_flutter/.flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"path_provider","path":"C:\\\\Users\\\\mdism\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider-1.6.24\\\\","dependencies":[]}],"android":[{"name":"path_provider","path":"C:\\\\Users\\\\mdism\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider-1.6.24\\\\","dependencies":[]}],"macos":[{"name":"file_chooser","path":"C:\\\\Users\\\\mdism\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\git\\\\flutter-desktop-embedding-040f36c82b63e8764ec2fd97066767503a667b6d\\\\plugins\\\\file_chooser\\\\","dependencies":[]},{"name":"path_provider_macos","path":"C:\\\\Users\\\\mdism\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_macos-0.0.4+6\\\\","dependencies":[]}],"linux":[{"name":"file_chooser","path":"C:\\\\Users\\\\mdism\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\git\\\\flutter-desktop-embedding-040f36c82b63e8764ec2fd97066767503a667b6d\\\\plugins\\\\file_chooser\\\\","dependencies":[]},{"name":"path_provider_linux","path":"C:\\\\Users\\\\mdism\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_linux-0.0.1+2\\\\","dependencies":[]}],"windows":[{"name":"file_chooser","path":"C:\\\\Users\\\\mdism\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\git\\\\flutter-desktop-embedding-040f36c82b63e8764ec2fd97066767503a667b6d\\\\plugins\\\\file_chooser\\\\","dependencies":[]},{"name":"path_provider_windows","path":"C:\\\\Users\\\\mdism\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_windows-0.0.4+3\\\\","dependencies":[]}],"web":[]},"dependencyGraph":[{"name":"file_chooser","dependencies":[]},{"name":"path_provider","dependencies":["path_provider_macos","path_provider_linux","path_provider_windows"]},{"name":"path_provider_linux","dependencies":[]},{"name":"path_provider_macos","dependencies":[]},{"name":"path_provider_windows","dependencies":[]}],"date_created":"2020-12-09 07:13:03.944796","version":"1.24.0-8.0.pre.361"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"path_provider","path":"C:\\\\Users\\\\murad\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider-1.6.24\\\\","dependencies":[]}],"android":[{"name":"path_provider","path":"C:\\\\Users\\\\murad\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider-1.6.24\\\\","dependencies":[]}],"macos":[{"name":"file_chooser","path":"C:\\\\Users\\\\murad\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\git\\\\flutter-desktop-embedding-040f36c82b63e8764ec2fd97066767503a667b6d\\\\plugins\\\\file_chooser\\\\","dependencies":[]},{"name":"path_provider_macos","path":"C:\\\\Users\\\\murad\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_macos-0.0.4+6\\\\","dependencies":[]}],"linux":[{"name":"file_chooser","path":"C:\\\\Users\\\\murad\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\git\\\\flutter-desktop-embedding-040f36c82b63e8764ec2fd97066767503a667b6d\\\\plugins\\\\file_chooser\\\\","dependencies":[]},{"name":"path_provider_linux","path":"C:\\\\Users\\\\murad\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_linux-0.0.1+2\\\\","dependencies":[]}],"windows":[{"name":"file_chooser","path":"C:\\\\Users\\\\murad\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\git\\\\flutter-desktop-embedding-040f36c82b63e8764ec2fd97066767503a667b6d\\\\plugins\\\\file_chooser\\\\","dependencies":[]},{"name":"path_provider_windows","path":"C:\\\\Users\\\\murad\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_windows-0.0.4+3\\\\","dependencies":[]}],"web":[]},"dependencyGraph":[{"name":"file_chooser","dependencies":[]},{"name":"path_provider","dependencies":["path_provider_macos","path_provider_linux","path_provider_windows"]},{"name":"path_provider_linux","dependencies":[]},{"name":"path_provider_macos","dependencies":[]},{"name":"path_provider_windows","dependencies":[]}],"date_created":"2020-12-10 07:39:29.010650","version":"1.25.0-5.0.pre.125"}
