+++ with_flutter/assets/model.json
@@ -1,5 +1,5 @@
 {
   "hello": {
-    "hello": "bello"
+    "helloIam1": "bello"
   }
 }
